requirements.txt
httpx
pendulum
plotext
pydantic
pytest
python-dotenv
